% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fubupackage-package.R
\docType{package}
\name{fubupackage-package}
\alias{fubupackage}
\alias{fubupackage-package}
\title{fubupackage: A dataframe plot}
\description{
What the package does (The package produces a line plot of a dataframe utilizing two functions).
}
\author{
\strong{Maintainer}: Ifebunandu Okeke \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
