% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Two level function of a combined dataframe}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{containing the first three numeric elements of this sequence}

\item{n}{denotes the final nth element of the sequence to calculate}
}
\value{
a combined function of x and n
}
\description{
Two level function of a combined dataframe
}
\examples{
myseq_n(x = c(2,3,3), n = 3)
myseq_n(x = c(2,4,3), n = 4)
myseq_n(x = c(2, 4, 3), n = 5)
myseq_n(x = c(2, 4, 3), n = 6)
myseq_n(x = c(2, 4, 3), n = 7)

}
