% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n_2.R
\name{myseq_n_2}
\alias{myseq_n_2}
\title{Dataframe line plot}
\usage{
myseq_n_2(df)
}
\arguments{
\item{df}{dataframe with four columns}
}
\value{
a line plot
}
\description{
Dataframe line plot
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)


myseq_n_2(my_data)

}
